feat: Saved canceled PDFs as files

+++ apps/invoicing/views/invoice.py
@@ -4,6 +4,7 @@ from itertools import chain
 
 from django.contrib.auth.decorators import login_required
 from django.contrib.auth.mixins import LoginRequiredMixin
+from django.core.files.base import ContentFile
 from django.db.models import DecimalField, ExpressionWrapper, F, Sum
 from django.http import HttpResponse
 from django.shortcuts import redirect, render
@@ -133,15 +134,22 @@ class InvoicePDFView(LoginRequiredMixin, DetailView):
     def get(self, request, *args, **kwargs):
         invoice = self.get_object()
 
-        file = generate_invoice(invoice, request)
+        if invoice.status == "CANCELED":
+            with open(invoice.pdf_file.path, "rb") as pdf:
+                response = HttpResponse(pdf.read(), content_type="application/pdf")
+                response["Content-Disposition"] = (
+                    f'filename="Invoice {invoice.id} - {invoice.matter} - {invoice.date_issued}.pdf"'
+                )
+        else:
+            file = generate_invoice(invoice, request)
 
-        with open(file.name, "rb") as pdf:
-            response = HttpResponse(pdf.read(), content_type="application/pdf")
-            response["Content-Disposition"] = (
-                f'filename="Invoice {invoice.id} - {invoice.matter} - {invoice.date_issued}.pdf"'
-            )
+            with open(file.name, "rb") as pdf:
+                response = HttpResponse(pdf.read(), content_type="application/pdf")
+                response["Content-Disposition"] = (
+                    f'filename="Invoice {invoice.id} - {invoice.matter} - {invoice.date_issued}.pdf"'
+                )
 
-        os.unlink(file.name)
+            os.unlink(file.name)
 
         return response
 
@@ -166,6 +174,16 @@ class StatusUpdateView(LoginRequiredMixin, View):
         invoice.save()
 
         if invoice_status == "CANCELED":
+            pdf_file = generate_invoice(invoice, request)
+
+            with open(pdf_file.name, "rb") as pdf:
+                invoice.pdf_file.save(
+                    f"Invoice {invoice.id} - {invoice.matter} - {invoice.date_issued}.pdf",
+                    ContentFile(pdf.read()),
+                )
+
+            invoice.save()
+
             TimeEntry.objects.filter(invoice=invoice).update(invoice=None)
             ExpenseEntry.objects.filter(invoice=invoice).update(invoice=None)
